# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/user": {
            "post": {
                "description": "creates a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "create a new user",
                "parameters": [
                    {
                        "description": "User creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/user/id/:username": {
            "get": {
                "description": "retrieves information about the user specified by the username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "get a user's info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the user to retrieve",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information",
                        "schema": {
                            "$ref": "#/definitions/api.UserGetResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid username",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes a user based on it's path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "deletes a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/login/gitlab/receiveToken": {
            "get": {
                "description": "where the user's browser is sent by GitLab after completing the oauth2 flow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "handles GitLab oauth2 callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The code returned by GitLab",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The state returned by GitLab",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GitLabReceiveTokenReponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.GitLabReceiveTokenReponse": {
            "type": "object"
        },
        "api.UserCreateRequest": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "boolean"
                },
                "displayName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.UserGetResponse": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}